if statement

+++ code/game/objects/items/implants/implantuplink.dm
@@ -28,6 +28,8 @@ CREATION_TEST_IGNORE_SUBTYPES(/obj/item/implant/uplink)
  */
 /obj/item/implant/uplink/proc/_component_removal(datum/source, datum/component/component)
 	SIGNAL_HANDLER
+	if(QDELETED(src))
+		return
 	if(istype(component, /datum/component/uplink))
 		qdel(src)
